+++ network-api/networkapi/mozfest/wagtail_hooks.py
@@ -1,13 +1,27 @@
 from wagtail.snippets.models import register_snippet
 from wagtail.snippets.views.snippets import SnippetViewSet, SnippetViewSetGroup
 
+from networkapi.events.models import TitoEvent
 from networkapi.mozfest.models import NewsletterSignupWithBackground, Ticket
+
+
+class TitoEventSnippetViewSet(SnippetViewSet):
+    model = TitoEvent
+    icon = "tito"
+    menu_order = 000
+    menu_label = "Tito"
+    menu_name = "Tito"
+    list_display = (
+        "title",
+        "event_id",
+    )
+    search_fields = ("title", "event_id", "newsletter_question_id")
 
 
 class TicketSnippetViewSet(SnippetViewSet):
     model = Ticket
     icon = "ticket"
-    menu_order = 000
+    menu_order = 100
     menu_label = "Tickets"
     menu_name = "Tickets"
     list_display = (
@@ -22,7 +36,7 @@ class TicketSnippetViewSet(SnippetViewSet):
 class NewsletterSignupWithBackgroundSnippetViewSet(SnippetViewSet):
     model = NewsletterSignupWithBackground
     icon = "newspaper"
-    menu_order = 100
+    menu_order = 200
     menu_label = "Newsletter Signups"
     menu_name = "Newsletter Signups"
     list_display = (
@@ -33,7 +47,7 @@ class NewsletterSignupWithBackgroundSnippetViewSet(SnippetViewSet):
 
 
 class MozfestViewSetGroup(SnippetViewSetGroup):
-    items = (TicketSnippetViewSet, NewsletterSignupWithBackgroundSnippetViewSet)
+    items = (TitoEventSnippetViewSet, TicketSnippetViewSet, NewsletterSignupWithBackgroundSnippetViewSet)
     menu_icon = "mozfest"
     menu_label = "Mozfest"
     menu_name = "Mozfest"
